added support for a second descriptor set with textures (Albedo, Normal, Physical(roughness and metal for now)) cleaned debug messages and added vulkan validation layer, can be turned on and of from the engine settings

+++ src/render/pipeline_infos.rs
@@ -14,5 +14,5 @@ pub struct Main {
 ///The definition of the "material textures"
 #[derive(Clone)]
 pub struct MaterialTextures {
-    pub albedo : f32,
+    pub albedo : i32,
 }
